- Move code originally in BuildLibraryNotification from SkylineWindow to PeptideSettingsUI - Stop using ModifyDocument during library build completion. Change the RT combo box selection instead.
